requirements.txt
pandas>=1.5.0
openpyxl>=3.1.0
datetime
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
pyodbc>=4.0.39
python-dotenv>=1.0.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-core>=1.29.5
requests>=2.31.0
msal>=1.24.0
tenacity>=8.2.3